Major rebuild and restructure, added conf file and daemon options otherwise functionality should be the same (get_attribute function is not implemented at the moment).
The 32/64 bit portability issue is handled better 
- at compile time, although I cannot guarantee 
this will work as intended on something exotic.

You will need to tweak the makefiles to put things 
where you want them and point to correct 
xmlrpc-c-config.

The default conf file is drmaa-xmlrpc.conf in cwd 
or can be overwritten via cli flag 
--conf-file=/path/to/file.

As of right now no log file is kept, I will be 
adding that at some point.

There are beginnings of a cgi possibility, but it 
will not work with current principle as the 
allocated job templates will go away between cgi 
calls